+++ CHANGELOG.md
@@ -1,10 +1,14 @@
-# [1.4.1](https://github.com/kenherring/ablunit-test-runner/releases/tag/1.4.1) - 2025-11-19 (pre-release)
+# [1.4.3](https://github.com/kenherring/ablunit-test-runner/releases/tag/1.4.3) - 2025-11-26 (pre-release)
 
+* Add `initializationProcedure` option to ablunit-test-profile.json (#453)
+* Bump typescript-eslint from 8.46.4 to 8.48.0 (#449)
+* Bump @vscode/vsce from 3.7.0 to 3.7.1 (#450)
 * Bump glob from 10.3.10 to 10.5.0 in /dummy-ext in the npm_and_yarn group across 1 directory (#447)
+* Bump @types/node from 24.10.0 to 24.10.1 (#443)
 * Bump the npm_and_yarn group across 1 directory with 1 update (#446)
 * Bump js-yaml from 3.14.1 to 3.14.2 in the npm_and_yarn group across 1 directory (#445)
 
-**Full Changelog**: [1.4.0...1.4.1](https://github.com/kenherring/ablunit-test-runner/compare/1.4.0...1.4.1)
+**Full Changelog**: [1.4.0...1.4.3](https://github.com/kenherring/ablunit-test-runner/compare/1.4.1...1.4.3)
 
 # [1.4.0](https://github.com/kenherring/ablunit-test-runner/releases/tag/1.4.0) - 2025-11-12